+++ NEWS.md
@@ -1,3 +1,11 @@
+Knet v1.2.8 Release Notes
+=========================
+
+* Fixed #506: RNN serialization and gc issues.
+* Fixed #509: Default output for progress is now stderr.
+* Added NCE chapter to docs.
+
+
 Knet v1.2.7 Release Notes
 =========================
 9fe1e03 2019-09-29